Refactor file structure and extracted JWT validation code in auth package
