+++ config.example.toml
@@ -1,5 +1,6 @@
 listen = "0.0.0.0:3000"
 base_url = "http://localhost:3000"
+client_urls = ["http://localhost:5173"]
 
 installation_id = ""
 app_id = ""
